requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
dj-database-url==0.4.2
Django==1.11.5
django-extensions==1.9.1
djangorestframework==3.6.4
elasticsearch==5.4.0
geoip2==2.5.0
idna==2.6
maxminddb==1.3.0
model-mommy==1.4.0
names==0.3.0
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
six==1.11.0
urllib3==1.22
